Update listing 18.11 (merge files and minor changes)
