+++ src/Chapter18/Listing18.12.csprojMockFileForResequencingListings.Placeholder.cs
